'Beyond Traditional Frameworks For Modelling Learning Domains'
Information and communications technology (ICT) is becoming increasingly
important in the educational sector. These technologies allow process
automation to take advantage of existing computational resources,
improving teaching experience. In this context, Machine Learning
provides an added value because it is capable of automatically modelling
complex problems that used to be solved using traditional methods. The
purpose of this thesis is to research the state of the art in Machine
Learning techniques to help teachers modelling and predicting students
performance. More specifically, by using Gradient Boosting, a highly
competent Ensemble method.